Added new function which finds all the words created on a specific board.
Input is the board which is to be analyzed by the function. Output is a vector of strings of the words in the entire board analyzed. This is used by the function countPoints.cpp on the original configuration of the board and on the new configuration after player tile placement. Comparing the words on the board in the original and new configurations can let us find all the new words created by a player during their turn, which is the main purpose for having this function.