Adds a function to execute a player's move
Inputs are an integer signalling whose turn it is, an integer represting number of players, a vector of strings of the letters left in the “bag”, a 2-dimensional vector of strings holding the current board information, and a dynamic vector of structs Player holding all the players’ information. Output is a boolean whether the player passed their turn or not. (This is necessary, because in the situation where no more tile placement is possible, players can each pass twice in a row to end the game early.)